[BTAT-2248] Moved NinoModel to models.core package
