+++ it/assets/NinoIntegrationTestConstants.scala
@@ -17,7 +17,8 @@
 package assets
 
 import assets.BaseIntegrationTestConstants.testNino
-import models.{NinoErrorModel, NinoModel}
+import models.NinoErrorModel
+import models.core.{NinoErrorModel, NinoModel}
 import play.api.libs.json.Json
 
 object NinoIntegrationTestConstants {
