+++ it/controllers/IncomeSourceDetailsControllerISpec.scala
@@ -21,8 +21,9 @@ import assets.IncomeSourceIntegrationTestConstants._
 import assets.NinoIntegrationTestConstants._
 import helpers.ComponentSpecBase
 import helpers.servicemocks.DesBusinessDetailsStub
+import models.core.{NinoErrorModel, NinoModel}
 import models.incomeSourceDetails.{IncomeSourceDetailsError, IncomeSourceDetailsModel}
-import models.{IncomeSourceDetailsError, NinoErrorModel, NinoModel}
+import models.{IncomeSourceDetailsError, NinoErrorModel}
 import play.api.http.Status._
 
 class IncomeSourceDetailsControllerISpec extends ComponentSpecBase {
