+++ test/assets/IncomeSourceDetailsTestConstants.scala
@@ -20,6 +20,7 @@ import assets.BaseTestConstants._
 import assets.BusinessDetailsTestConstants._
 import assets.PropertyDetailsTestConstants._
 import models._
+import models.core.{NinoErrorModel, NinoModel}
 import models.incomeSourceDetails.{IncomeSourceDetailsError, IncomeSourceDetailsModel}
 import play.api.libs.json.Json
 import play.mvc.Http.Status
